Fixes #15 - per entity controllers
