+++ app/Http/Controllers/Controller.php
@@ -2,12 +2,21 @@
 
 namespace App\Http\Controllers;
 
+use App\Project;
 use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Illuminate\Foundation\Bus\DispatchesJobs;
 use Illuminate\Foundation\Validation\ValidatesRequests;
 use Illuminate\Routing\Controller as BaseController;
+use Illuminate\Support\Facades\View;
 
 class Controller extends BaseController
 {
     use AuthorizesRequests, DispatchesJobs, ValidatesRequests;
+
+    public function __construct()
+    {
+        $languages = Project::distinct()->get('language');
+
+        View::share('all_languages', $languages);
+    }
 }
